pyproject.toml
[tool.poetry]
name = "sharky-bot"
version = "0.1.0"
description = "sharky-bot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.6"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-remake = "^0.2.7"
requests = "^2.28.1"
nonebot-plugin-gocqhttp = "^0.6.3"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[tool.nonebot]
plugins = ["nonebot_plugin_remake", "nonebot_plugin_gocqhttp"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "bfsu"
default = true
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple"